Add load_from_file to the main API.
Explicit is better than implicit.  Previously load_from would try to magically
guess whether the string you gave it was a filename or actual content.  Not only
was this error-prone (is "hello world" a filename or a single subtitle?) but it
would actually crash when passed tricky unicode data.

It would also open() the file, do some stuff, and then use it in a with block,
which means if it crashed during the "do some stuff" phase the file would never
get closed.  That's fixed here too.