Refactored SubtitleSet a bit
Store the body element when when we initialize our TTML.  We use it a
lot and this makes it simpler/faster to lookup.  Added some
convenience methdos for finding our DIVs.

Changed how we create our <p> in append_subtitle().  I think the new
method is easier to understand and a bit cleaner.  Also changed how we
create the <div> when new_paragraph is True.